refactor: implement weekSelector component to replace select menu in ActivityDetailsCard and SummaryTitle
